% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate2adm.R
\name{aggregate2adm}
\alias{aggregate2adm}
\title{Aggregate raster files to adm level}
\usage{
aggregate2adm(r, adm)
}
\arguments{
\item{r}{Object of class SpatRaster or Raster}

\item{adm_file}{Object of class sf with location of subnational administrative units}
}
\value{

}
\description{
Function to aggregate raster files to adm level.
}
