% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipu2.R
\name{check_tables}
\alias{check_tables}
\title{Check seed and target tables for completeness}
\usage{
check_tables(
  primary_seed,
  primary_targets,
  secondary_seed = NULL,
  secondary_targets = NULL,
  primary_id
)
}
\value{
both seed tables and target lists
}
\description{
Given seed and targets, checks to make sure that at least one
observation of each marginal category exists in the seed table.  Otherwise,
ipf/ipu would produce wrong answers without throwing errors.
}
\keyword{internal}
