% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip_country.R
\name{clip_country}
\alias{clip_country}
\title{Function to clip national maps from global WorldPop sex and age disaggregated population maps}
\usage{
clip_country(r, adm)
}
\arguments{
\item{r}{Object of class SpatRaster, Raster or their path}

\item{adm}{object of class sf with the location of subnational units}
}
\value{

}
\description{
Function to clip national maps from global WorldPop sex and age disaggregated population maps
}
