% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_folders.R
\name{create_folders}
\alias{create_folders}
\title{Creates \code{ssid} folder structure}
\usage{
create_folders(param = NULL)
}
\arguments{
\item{param}{Object of type \code{ssid_par} that bundles all \code{ssid} parameters,
including core model folders, alpha-3 country code and year and administrative unit
level at which the model is solved and type of
model.}
}
\description{
\code{create_folders} creates the folder structure that is needed to store the processed data for
\code{ssid}.
}
\details{
\code{create_folders} creates two folders in the \code{model_path}, set by the user:
mappings and processed_data (including subfolders). It copies a number of cvs files
into the mappings folder, which contain several data tables that are needed to run
the model and, if needed can be adjusted by the user.
}
\examples{
\dontrun{
create_folders(param)
}

}
