% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_template.R
\name{create_model_template}
\alias{create_model_template}
\title{Creates \code{ssid} model template}
\usage{
create_model_template(template_path = NULL)
}
\arguments{
\item{template_path}{folder where template scripts will be copied.}
}
\description{
\code{create_model_template} creates a folder structure with template R scripts to implement the various
steps to run \code{ssid}.
}
\details{
\verb{create_model template()} creates several folders in the \code{template_path}, set by the user. The folders
contain template R scripts to implement \code{ssid}.

To run the template scripts, they have to be copied into an RStudio project. We recommend to
create such a project first and then use \code{create_model_template()} to add the scripts to
the RStudio project. Note that \code{create_model_template()} only copies script files if they do not exist
yet. In this way, the user cannot accidentally replace already modified scripts.
}
\examples{
\dontrun{
create_model_template("c:/temp")
}

}
