% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_weights_plot.R
\name{create_weights_plot}
\alias{create_weights_plot}
\title{Title
Create figure with weight distribution}
\usage{
create_weights_plot(region, sim_file)
}
\arguments{
\item{region}{Region for which weights are extracted}

\item{sim_file}{Simulation file}
}
\description{
Title
Create figure with weight distribution
}
