% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_output.R
\name{extract_output}
\alias{extract_output}
\title{Extract weights and diagnostics from spatial simulation object}
\usage{
extract_output(sim, ssp_y, reg_tz, output)
}
\arguments{
\item{sim}{Object with spatial simulation results.}

\item{ssp_y}{Vector with ssp-year combinations for which simulation results are stored in \code{sim}.}

\item{reg_tz}{Vector with regions for which simulation results are stored in \code{sim}.}

\item{output}{Type of output that needs to be extracted.}
}
\description{
Extract weights and diagnostics from spatial simulation object
}
