% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_weight_plot.R
\name{extract_weight_plots}
\alias{extract_weight_plots}
\title{Extract figures with with weight distribution}
\usage{
extract_weight_plots(sim, ssp_y, reg_tz)
}
\arguments{
\item{sim}{Object with spatial simulation results.}

\item{ssp_y}{Vector with ssp-year combinations for which simulation results are stored in \code{sim}.}

\item{reg_tz}{Vector with regions for which simulation results are stored in \code{sim}.}
}
\description{
Extract figures with with weight distribution
}
