% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_outlier.R
\name{fix_outlier_quantile}
\alias{fix_outlier_quantile}
\title{Function to fix outliers using quantile approach}
\usage{
fix_outlier_quantile(x, q_lower = 0.025, q_upper = 0.975)
}
\arguments{
\item{x}{Vector with data containing outliers}

\item{q_lower}{Lower quantile range. Default is 0.025.}

\item{q_upper}{Upper quantile range. Default is 0.975.}
}
\description{
This function uses quantiles to determine outliers and replace them with the median value.
}
\examples{
fix_outlier_quantile(rnorm(40, mean = 0, sd = 1))
}
