% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf_seed.R
\name{ipf_seed}
\alias{ipf_seed}
\title{ipf_seed}
\usage{
ipf_seed(
  reg_tz,
  hh_s,
  per_s,
  bm,
  p,
  reg_sample = FALSE,
  verbose = FALSE,
  max_iter = 300,
  max_ratio = 5,
  min_ratio = 0.2,
  relative_gap = 0.01,
  absolute_diff = 10
)
}
\arguments{
\item{absolute_diff}{}
}
\value{

}
\description{
ipf_seed
}
