% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipu_subnat_proj.R
\name{ipu_subnat_proj}
\alias{ipu_subnat_proj}
\title{ipu_subnat_proj}
\usage{
ipu_subnat_proj(y, scen, verbose = TRUE)
}
\arguments{
\item{y}{Year}

\item{scen}{Scenario}

\item{verbose}{Print iteration details and worst marginal stats upon completion generated by \code{ipfr::ipu_matrix}?
Default TRUE.}
}
\value{

}
\description{
Function to rake subnational population projections so that they match national age-sex and
subnational pop projections.
}
\details{

}
