% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssid_par.R
\name{ssid_par}
\alias{ssid_par}
\title{Sets \code{ssid} parameters}
\usage{
ssid_par(
  model_path = NULL,
  db_path = NULL,
  iso3c = NULL,
  adm_level = NULL,
  base_year = NULL,
  start_year = NULL,
  end_year = NULL
)
}
\arguments{
\item{model_path}{character. Folder where the model files are stored. Note that R uses
forward slash or double backslash to separate folder names.}

\item{db_path}{character. Folder where the input database is stored.}

\item{iso3c}{character. Three letter ISO 3166-1 alpha-3 country
code, also referred to as iso3c. A list of country codes can be found in
\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}{Wikipedia}.}

\item{adm_level}{numeric. Administrative unit level at which the model is run. This needs to correspond with
the information in the shapefile of the administrative units, which is needed as input.}

\item{base_year}{numeric. Base year of the model, normally the year of the household survey.}

\item{start_year}{numeric. Start year of the simulation period.}

\item{end_year}{numeric. End year of the simulation period.}
}
\value{
ssid_par object
}
\description{
\code{ssid_par} sets all required parameters to run \code{ssid}, including core model
folders, country code and year.
}
\details{
\code{ssid_par} creates an object of class \code{ssid_par}, which bundles all required
\code{ssid} parameters set by the user: model and database folder, three digit country code, country name and
continent. The coordinate reference system is automatically set to WGS84
(epsg:4326).

\code{\link[countrycode:countrycode]{countrycode::countrycode()}} is used to determine the full country
name, three digit country code and continent on
the basis of the alpha-3 country code. This information is required to extract
country specific information from several datasets.
}
\examples{
\dontrun{
ssid_par(model_path = "C:/temp/ssid_eth", db_path = "C:/temp/ssid_db",
iso3c = "ETH", adm_level = 2, base_year = 2018, start_year = 2018, end_year = 2050)
}
}
