% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_output.R
\name{extract_output}
\alias{extract_output}
\title{Extract weights and diagnostics from spatial simulation object}
\usage{
extract_output(df, var)
}
\arguments{
\item{df}{Object with spatial simulation results nested by ssp-year and
region.}

\item{var}{Type of output that needs to be extracted. Options are: stats,
weight_tbl, primary_comp, secondary_comp}
}
\description{
Extract weights and diagnostics from spatial simulation object
}
