% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reweigh.R
\name{reweigh}
\alias{reweigh}
\title{Title}
\usage{
reweigh(
  ssp_y,
  reg_tz,
  hh_s,
  per_s,
  bm,
  verbose = FALSE,
  reg_sample = FALSE,
  max_iter = 300,
  max_ratio = 5,
  min_ratio = 0.2,
  relative_gap = 0.01,
  absolute_diff = 10
)
}
\arguments{
\item{absolute_diff}{}
}
\value{

}
\description{
Title
}
